requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.3
psycopg==3.2.4
psycopg-binary==3.2.4
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
